# api/api-spec.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "WasteTrack API",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "description": "API for waste management and tracking system",
    "version": "1.0.0",
    "contact": {
      "name": "WasteTrack Team"
    }
  },
  "servers": [
    {
      "url": "https://api.wastetrack.com",
      "description": "Production server"
    },
    {
      "url": "https://staging-api.wastetrack.com",
      "description": "Staging server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/waste-transfer-requests/{id}": {
      "get": {
        "operationId": "GetWasteTransferRequest",
        "tags": ["User - Waste Transfer Requests"],
        "summary": "Get waste transfer request by ID",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Waste transfer request retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/WasteTransferRequestResponse"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/api/waste-bank/waste-transfer-requests/{id}/assign-collector": {
      "put": {
        "operationId": "AssignCollectorByWasteType",
        "tags": ["Waste Bank - Waste Transfer Requests"],
        "summary": "Assign collector and accept waste types",
        "description": "Assigns a collector to the waste transfer request and sets accepted weights and prices for each waste type.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Waste transfer request ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignCollectorByWasteTypeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Collector assigned successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WasteTransferRequestResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
    "/waste-transfer/{id}/complete": {
      "put": {
        "operationId": "CompleteWasteTransferRequest",
        "tags": ["Waste Bank - Waste Transfer Requests"],
        "summary": "Complete waste transfer request",
        "description": "Completes a waste transfer request by recording verified weights. This operation updates storage inventories and user profiles.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "Waste transfer request ID",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompleteWasteTransferRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Waste transfer completed successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WasteTransferRequestResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Missing required fields or invalid data",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Field validation failed"
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized - Not logged in",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Unauthorized access"
                }
              }
            }
          }
        }
      },
      "Forbidden": {
        "description": "Forbidden - Lacking permission",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Insufficient permissions"
                }
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - ID not found",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Resource not found"
                }
              }
            }
          }
        }
      },
...
      "InternalServerError": {
        "description": "Internal Server Error - Server error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "errors": {
                  "type": "string",
                  "example": "Internal server error"
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
...
      "LocationResponse": {
        "type": "object",
        "properties": {
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        },
        "required": ["latitude", "longitude"]
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "username": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string"
          },
          "avatar_url": {
            "type": "string"
          },
          "phone_number": {
            "type": "string"
          },
          "institution": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "province": {
            "type": "string"
          },
          "points": {
            "type": "integer",
            "format": "int64"
          },
          "balance": {
            "type": "integer",
            "format": "int64"
          },
          "location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "is_email_verified": {
            "type": "boolean"
          },
          "is_accepting_customer": {
            "type": "boolean"
          },
          "access_token": {
            "type": "string"
          },
          "refresh_token": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "example": "2023-08-02T10:00:00.000Z"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-08-02T10:00:00.000Z"
          },
          "is_agreed_to_terms": {
            "type": "boolean"
          },
          "distance": {
            "type": "number",
            "format": "float"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "role",
          "points",
          "balance",
          "is_email_verified"
        ]
      },
...
      "WasteCategoryResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": ["id", "name", "description"]
      },
...
      "WasteTypeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "category_id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "waste_category": {
            "$ref": "#/components/schemas/WasteCategoryResponse"
          }
        },
        "required": ["id", "category_id", "name", "description"]
      },
...
      "WasteTransferItemOfferingResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "transfer_form_id": {
            "type": "string"
          },
          "waste_type_id": {
            "type": "string"
          },
          "offering_weight": {
            "type": "number",
            "format": "float"
          },
          "offering_price_per_kgs": {
            "type": "number",
            "format": "float"
          },
          "accepted_weight": {
            "type": "number",
            "format": "float"
          },
          "accepted_price_per_kgs": {
            "type": "number",
            "format": "float"
          },
          "transfer_form": {
            "$ref": "#/components/schemas/WasteTransferRequestSimpleResponse"
          },
          "waste_type": {
            "$ref": "#/components/schemas/WasteTypeResponse"
          }
        }
      },
...
      "WasteTransferRequestSimpleResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "source_user_id": {
            "type": "string"
          },
          "destination_user_id": {
            "type": "string"
          },
          "form_type": {
            "type": "string"
          },
          "total_weight": {
            "type": "integer"
          },
          "total_price": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "source_phone_number": {
            "type": "string"
          },
          "destination_phone_number": {
            "type": "string"
          },
          "appointment_date": {
            "type": "string",
            "format": "date"
          },
          "appointment_start_time": {
            "type": "string"
          },
          "appointment_end_time": {
            "type": "string"
          },
          "appointment_location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "distance": {
            "type": "number",
            "format": "float",
            "example": 0
          }
        }
      },
      "WasteTransferRequestResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "source_user_id": {
            "type": "string"
          },
          "destination_user_id": {
            "type": "string"
          },
          "form_type": {
            "type": "string"
          },
          "total_weight": {
            "type": "integer"
          },
          "total_price": {
            "type": "integer"
          },
          "status": {
            "type": "string"
          },
          "image_url": {
            "type": "string"
          },
          "notes": {
            "type": "string"
          },
          "source_phone_number": {
            "type": "string"
          },
          "destination_phone_number": {
            "type": "string"
          },
          "appointment_date": {
            "type": "string",
            "format": "date"
          },
          "appointment_start_time": {
            "type": "string"
          },
          "appointment_end_time": {
            "type": "string"
          },
          "appointment_location": {
            "$ref": "#/components/schemas/LocationResponse"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          },
          "source_user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "destination_user": {
            "$ref": "#/components/schemas/UserResponse"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WasteTransferItemOfferingResponse"
            }
          },
          "distance": {
            "type": "number",
            "format": "float",
            "example": 0
          }
        }
      },
...
      "AssignCollectorByWasteTypeRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "assigned_collector_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the collector to assign (optional - some industries can collect without specific collector)",
            "example": "123e4567-e89b-12d3-a456-426614174000"
          },
          "items": {
            "type": "array",
            "description": "Waste types with accepted weights and prices",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/AssignCollectorWasteTypeRequest"
            }
          }
        }
      },
      "AssignCollectorWasteTypeRequest": {
        "type": "object",
        "required": [
          "waste_type_id",
          "accepted_weight",
          "accepted_price_per_kgs"
        ],
        "properties": {
          "waste_type_id": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the waste type",
            "example": "456e7890-e89b-12d3-a456-426614174001"
          },
          "accepted_weight": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "description": "Accepted weight in kilograms",
            "example": 50.5
          },
          "accepted_price_per_kgs": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "description": "Accepted price per kilogram",
            "example": 2.5
          }
        }
      },
      "CompleteWasteTransferRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "$ref": "#/components/schemas/CompleteWasteTransferRequestItems"
          }
        }
      },
      "CompleteWasteTransferRequestItems": {
        "type": "object",
        "required": ["waste_type_ids", "weights"],
        "properties": {
          "waste_type_ids": {
            "type": "array",
            "description": "Array of waste type IDs",
            "minItems": 1,
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "example": [
              "456e7890-e89b-12d3-a456-426614174001",
              "789e0123-e89b-12d3-a456-426614174002"
            ]
          },
          "weights": {
            "type": "array",
            "description": "Array of verified weights in kilograms (must match waste_type_ids array length)",
            "minItems": 1,
            "items": {
              "type": "number",
              "format": "float",
              "minimum": 0
            },
            "example": [48.2, 32.8]
          }
        }
      },
...
    }
  }
}